+++ DOCS/CHANGELOG.md
@@ -16,6 +16,7 @@
   - work: `ExtProcess`.
 - new consts:
   - media::font: `FONT_3_3`, `FONT_3_5`, `FONT_5_6`.
+  - `Float`|`ExtFloatConst`: `EXPONENT_BIAS`, `EXPONENT_BITS`, `SIGNIFICANT_BITS`.
   - `AngleDirection::{CounterClockwise, CCW, RightHandRule, RHR, Clockwise, CW, LeftHandRule, LHR}`.
 - new types:
   - code: `ScopeGuard`.